app/requirements.txt
aiofiles==23.1.0
alembic==1.10.2
anyio==3.6.2
attrs==22.2.0
bitarray==2.7.3
black==23.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.95.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.2.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pydantic==1.10.7
pytest==7.2.2
python-multipart==0.0.6
requests==2.28.2
sniffio==1.3.0
SQLAlchemy==2.0.7
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
